fix(pa): Pass decode output dtype explicitly
Made-with: Cursor

+++ op_tests/triton_tests/test_pa_decode_gluon.py
@@ -1545,7 +1545,7 @@ def run_pa_gluon_test(
     temporary_output = torch.empty(
         *intermediate_shape,
         head_size,
-        dtype=torch.float32,
+        dtype=reference_output_quant.dtype,
         device=reference_output_quant.device,
     )
     # Create output tensor with the same shape as reference
